UI, shadows fixed

+++ Assets/_Scripts/UI/Buttons/Shop Buttons/UpgradeButton.cs
@@ -20,7 +20,7 @@ namespace _Scripts.UI.Buttons.Shop_Buttons
         protected override bool CanBeBought => Coefficient < maxUpgrade;
         #endregion
         
-        protected override void ChangeButtonState(int moneyCount)
+        protected override void ChangeButtonState(float moneyCount)
         {
             if (Coefficient >= maxUpgrade)
             {
